Upgrade to the last version of the `googlesheets` package.
There have been several changes in the package, as it is in development
state. For example, the function `register_ss ( )` has been replaced by
the function `gs_title ( )`. I'll have to keep an eye for future changes
like that.

The second addressed issue has to do with Windows and its encoding hell.
The scripts run smoothly in Mac OS X, but strange things happened when I
tried to run the code in Windows 7.

Slight changes have been made to load the data using the proper encoding
(UTF-8). Rstudio does not seem to show the data with the proper encoding
in the data tabs, but the data shows without flaws in the console, so I
guess this is more of an Rstudio issue.