This is the first major update in a long time, labeled as v0.91.
The most notorious change is the removal of the `Is.Shared` column. This column has not been used in a long time, so it didn't make any sense to keep it.

Apart from that, I have scanned all the source code files in search of typos and needed updates, tidying up a bit the project.